Modified Yay Install

+++ user_install.sh
@@ -2,12 +2,17 @@
 
 echo "Starting user-specific installation tasks..."
 
-# User-specific operations, for example:
-mkdir -p ~/github && cd ~/github
-git clone https://aur.archlinux.org/yay.git
-cd yay || exit
-makepkg -si --noconfirm
-cd .. && rm -rf yay
+#### Check for yay ####
+ISYAY=/sbin/yay
+if [ -f "$ISYAY" ]; then 
+    echo -e "yay was located, moving on."
+else 
+    echo -e "Yay was NOT located installing"
+    git clone https://aur.archlinux.org/yay.git &>> $INSTLOG
+    cd yay
+    makepkg -si --noconfirm &>> ../$INSTLOG
+    cd .. && rm -rf yay
 
+fi
 echo "User-specific installations complete."
